requirements.txt
httpx
telethon
python-dotenv
gspread
google-auth
google-auth-oauthlib
google-auth-httplib2
cryptography
